View Assignment put more safe guards in and added 1D Left/Right from Strided
This adds the possibility to assign a 1D strided view to a layoutleft or
LayoutRight view if its stride is 1.

It also guards the default assignment to strided view if it is not the same
layout to only be defined from LayoutLeft and LayoutRight. Other Layouts may
not be assignable to strided view (e.g. custom layouts).