Makefile system: adding LDFLAGS overload possibility
This turns out to be necessary to point LLVM to the correct clibs.
Clang did try to link against GCC 4.4.x on RH even after setting the
toolchain with --gcc-toolchain. The LDPath to the newer gcc had to be
explicitly set:
~/kokkos/generate_makefile.bash --with-pthread --compiler=clang++
  --kokkos-path=${HOME}/kokkos
  --cxxflags="-O3 --gcc-toolchain=/home/projects/x86-64/gcc/4.9.2"
  --ldflags="-L/home/projects/x86-64/gcc/4.9.2/lib64"

This worked for me with clang version
clang version 3.6.0 (trunk 221174) (llvm/trunk 221170)

on RH 6.5 with gcc 4.9.2 loaded.