Fixed issue of hanging atomics test with LLVM
The hanging test was atomic CAS on Kokkos::complex<double>.
This function is only atomic when X86 assembly is enabled, which
is not the case with LLVM. Using atomics on 128 bit values is
currently only valid with GCC and Intel compilers on the host.
When using Serial execution space atomics on complex<double>
are allowed always. On Cuda atomicAdd is allowed on complex<double>,
but it internally does two atomic adds (one on each member).

The test will now not excersice Kokkos::complex<double> atomics
when assembly for X86 is not enabled and one is using a threaded
execution space.