Finished initial implementation for arbitrary sizes atomic operations
This implements atomic operations for arbitrarily sized objects.
The implementation uses a hash table approach, where a lock is set
based on a hash of the memory address of the object for which an
atomic operation should be performed.
Initial performance results indicate that it is comparable to
other non-native atomics (i.e. CAS loops with casting to integer
types).
The commit implements the full set of supported atomics from
Kokkos and it works in all currently existing execution spaces.
The hashtables are static sized global arrays.

Note: this commit requires relocatable-device-code being
enabled for Cuda.