Enabled relocatable device code always.
This requires that the kokkos library is build as a static library.
The feature allows to compile device functions in one compilation unit
and than call it in another.

This is a very helpful feature to implement arbitrary atomics.
It will also (going forward) be required for some stuff in the task
based policies. It will also allow libraries of device functions.