Rework of arbitrary atomics on Cuda.
After figuring out that Trilinos linking breaks with relocatable device
code enabled in unrelated parts (e.g. epetra linking breaks), this
reworks the internal workings to not use a static sized global array
for the lock array on cuda but an allocation, whose pointer is
put into constant cache. This required some reworking internally.
In particular the symbol for the lock array has to live in files
which are guaranteed to be part of any kernel (i.e. CudaExec), and
the functions using that symbol have to be in the correct inclusion order
with that symbol.