Fixed an issue with managing the lock array on Cuda
The previous model caused an issue when doing a cycle of:

Cuda::initialize()
Cuda::finalize()
Cuda::initialize()

This fixed that. Note that this is not officially supported.
You are not officially allowed to call intialize after finalize
was called. It just happened to work more or less before, and
it is a situation which is encountered in some unit tests of
Trilinos. This commit makes them not break anymore.