Profiler: adding basic inbuild kernel timing mechanism
This commit adds the beginnings for profiling support in Kokkos.
Particular this adds hooks for profilers to collect data on a per kernel base.
These hooks are currently only added to the parallel_### (std::string , ...)
variants of kernel dispatch. The hooks get two std::strings, the provided
kernel name as well as the execution space (the latter is a stub right now).

The basic inbuild profiler will simply time the kernel and keep track of
minimum, maximum and average time as well as number of calls.
The profiling will also fence kernels before and after the kernel call.

This capability is activated by defining KOKKOS_ENABLE_PROFILING_COLLECT_KERNEL_DATA=1.

All functions associated with this are in the KokkosP::Experimental namespace.

I need this feature to start analysing MueLu and Nalu.