Profiler: Reworking the implementation logic a bit.
This prepares the thing for CMake by eliminating dependencies on
KokkosCore in the Basic Profiler implementation.

A Kokkos_Profiler.hpp/.cpp file are added which have the public
Kokkos API for steering/setup of the profiler. The functions in that
are responsible to call the functions of the actual profiler and provide
it with the correct strings/info such as kernel name.

The public functions of the Basic Profiler are the ones which
third party tools are supposed to replace.