Kokkos Profiling: renamed files and moved to impl
The profiling interface is not meant to be public (i.e. users should not
call it directly) thus it should probably live in impl.
We might need to disucss that again.
Also since the internal interface is meant as a permanent always on
capabiltiy of Kokkos I renamed the file in the Kokkos naming scheme.