test_all_sandia: Add capability to run jobs in parallel
Instead of leveraging Jenkins to achieve parallelism, put parallelism
directly in the script. It's simpler that way and also allows people
to take advantge of this outside of Jenkins.

Default will be to run eight jobs at once. CUDA jobs will NOT be
run in parallel.

Also, adds a test mode for the script that sleeps instead of running
Kokkos.

Much more graceful handling of missing modules.