Kokkos: Allow to build as part of stand-alone Teuchos
This commit allows Kokkos to build as a package outside of Trilinos under
another TriBITS-based project.  The initial target in the Teuchos stand-alone
build which I tested this with.  You just clone Kokkos under Teuchos and you
you configure and build.

I moved the FindTPL<TPLNAME>.cmake modules for CUSPARSE, HWLOC, Pthread, and
QTHREAD from Trilinos to kokkos/cmake/tpls/.  This allows a TriBITS project,
like stand-alone Teuchos, to just list these TPLs for Kokkos and done.

This example shows a weakness of the current TriBITS design.  It seems that a
TriBITS package should be able to automatically inject its TPLs into a parent
project.  However, it would be very hard to get the order of the TPLs between
packages correct.  The explicit list at the repo level does this.  Anyway, not
too bad

Fixes #117