Kokkos{Core,Containers}: Add operator*(std::complex,Kokkos::complex)
Kokkos::complex is a drop-in replacement for std::complex that works
in Kokkos parallel kernels, both because it adds the __device__
attribute where appropriate, and because it has volatile overloads for
certain methods.

I noticed some build errors recently, because Tpetra was trying to
compute an expression like this:

const std::complex<double> alpha = ...;
Kokkos::complex<double> y_i = y(i);
y_i += alpha * A(i,j) * x(j);

where y(i), A(i,j), and x(j) are Kokkos::complex<double>.  The issue
was the expression "alpha * A(i,j)".  This only showed up now, because
before, Tpetra would always cast alpha to Kokkos::complex before using
it in the above expression.

The way I fixed this was to add a new nonmember operator* function:

template<class T>
Kokkos::complex<T>
operator* (const std::complex<T>&, const Kokkos::complex<T>&);

Note that this is NOT marked with the KOKKOS_INLINE_FUNCTION macro.
It can't be, because std::complex::real and std::complex::imag (which
this function must use) are not marked as CUDA device functions.
(That's the whole reason why Kokkos::complex exists.)

NOTE: I made these changes to the version of Kokkos in Trilinos before,
but NOT to the "master" version of Kokkos on Github.  The latter is the
right place to make Kokkos changes.  Changes to the version of Kokkos
in Trilinos get clobbered by periodic snapshots from the Github version
of Kokkos.