KokkosCore: Add missing Kokkos::complex<T> op*(T, Kokkos::complex<T>)
Kokkos::complex was missing the following templated nonmember function:

template<class T>
Kokkos::complex<T> operator* (const T&, const Kokkos::complex<T>&);

This commit adds it.

NOTE: I made these changes to the version of Kokkos in Trilinos before,
but NOT to the "master" version of Kokkos on Github.  The latter is the
right place to make Kokkos changes.  Changes to the version of Kokkos
in Trilinos get clobbered by periodic snapshots from the Github version
of Kokkos.