Cuda: fix for bug issue #125
Fixes an issue with cuda_get_max_block_size and cuda_get_opt_block_size.
This makes the choice of constant vs local memory a template parameter
defaulted by the size of the existing DriverType template parameter.
It also changes the interface by adding a new shmem_extra argument which is
required for lambdas since the functor in those cases doesn't have a
shmem size function.

Both functions are part of the impl namespace and thus not public yet.