nvcc_wrapper: fixed an issue if no host flag is given
The recent update to nvcc wrapper broke it if not a single host
compiler flag is given (such as -mavx). This fixes that.
This was not caught right away because: (i) most builds did set one
(for example through specifying a host architecture), and (ii) the
nightly testing was done with an installed nvcc_wrapper instead of the
one in the repo.