Kokkos::Experimental View - fix unmanaged views so that they do not reference count.
A View may be compile-time unmanaged with the MemoryUnmanaged trait.
A View may be run-time unmanaged if it is constructed from a pointer or assigned from an unmanaged view.

As View::use_count() method to query the reference count of any View
that was initially allocated by Kokkos, including currently unmanaged Views.