Kokkos: Check for oversubscription of CPUs
This adds a check for oversubscription of CPUs with too many threads
which also works in most common MPI environments. In particular it
works with OpenMPI, MVAPICH2 as well as with SLURM dispatch. It was
tested on Cray.

It checks that the per node MPI ranks times the number of threads per
process does not exceed the total number of cores (or hyperthreads).

This check happens for both OpenMP and Pthreads initialization.

This addesses issue #149.