Core: Workaround for clang with template matching issues.
Effectively at some point it tries to instantiate FunctorValueTraits<RangePolicy>
in a situation where it will clearly fail later to match, sicne the other function
arguments don't match. (instantiation of parallel_reduce). The problem is that
FunctorValueTraits expects the incoming thing to have an operator(), and if it
doesn't it produces a compiler error.