Fixing: memory_pool for OpenMP and Pthreads enabled at the same time
This was a compilation issue, triggered by the fact that the memory pool
was calling a parallel for but was only templated on memory space.
With OpenMP enabled that means that the Threads task policy was trying to start
an OpenMP dispatch, wihtout even including the OpenMP header files.

Not sure why this was not caught by the nightlies.