MemoryPool: fix installation issue
This addresses a problem when using kokkos in an installed setting (stand alone or with Trilinos),
with Cuda enabled. To work around a bug with respect to relocatable device code in nvcc,
Kokkos_MemoryPool.hpp included impl/Kokkos_MemoryPool.cpp. generally installing cpp files is not
part of the build systems. This broke compilation of applications like Albany against an installed
Trilinos when Cuda is enabled. The fix moves the code from the cpp file into a hpp file which
is subsequently included. The cpp file includes that hpp file.