First TaskPolicy unit tests working for CUDA.
Add '--with-cuda-options=*' to generate_makefile.bash.

Use the KOKKOS_CUDA_USE_RELOCATABLE_DEVICE_CODE macro
to control compilation of the MemoryPool code in
the 'cpp' file instead of inlining for CUDA builds.

Revise TaskPolicy API to work around CUDA limitation.
With CUDA the TaskPolicy::create and TaskPolicy::create_team
member functions cannot be called both within a task
and by the main process.
Split these into
	TaskPolicy::task_create
	TaskPolicy::task_create_team
	TaskPolicy::proc_create
	TaskPolicy::proc_create_team
which can be called from within a task or by the main process
as per the 'task' and 'proc' qualifier.

Implement and test Future< Latch > for both Cuda and Threads.

Cuda and Threads TaskPolicy to increment task reference count when spawned
and decrement when complete.  Required to properly handle
"spawn and forget" tasks.