Core: Bugfix subviews RandomAccess
This fixes #385 and #310. To do so I had to change internal
interface for assigning data handles. This might affect
Stokhos/Sacado. Basically the subview assignment has to
give the handle and the offset as two separate things.
Before it was using raw pointers with offset added directly.
For handle types which need to keep offset as its own thing
(like Cuda Texture objects) this is not possible.