Disable GTest pthread use if Pthread not enabled
Not sure where the best location for this test is, but if TPL_ENABLE_Pthread is not enabled, then Gtest has to be told to not enable Pthread use also.  Currently, the following code in gtest.h will enable pthreads if GTEST_HAS_PTHREAD is not defined externally:

```
// Determines whether Google Test can use the pthreads library.
#ifndef GTEST_HAS_PTHREAD
// The user didn't tell us explicitly, so we assume pthreads support is
// available on Linux and Mac.
//
// To disable threading support in Google Test, add -DGTEST_HAS_PTHREAD=0
// to your compiler flags.
# define GTEST_HAS_PTHREAD (GTEST_OS_LINUX || GTEST_OS_MAC || GTEST_OS_HPUX \
    || GTEST_OS_QNX)
#endif  // GTEST_HAS_PTHREAD
```

This will enable gtest pthread use even if TPL_ENABLE_Pthread is disabled which can result in undefined symbols when linking.

I may be missing some other configuration setting that handles this.