CudaSpace: Add tracking for num of UVM allocs
This refers to Kokkos Github issue #300.
There is a limit of 64k i.e. 65536 Cuda UVM allocations after which an
"out of memory" error is issue.
This commit adds tracking of the number of allocations to CudaUVM space
and throws a runtime error when the max number of UVM allocations is
reached, along with a message for users.

	modified:   core/src/Cuda/Kokkos_CudaSpace.cpp
	modified:   core/src/Kokkos_CudaSpace.hpp
	modified:   core/unit_test/cuda/TestCuda_Spaces.cpp