Core: Fix issue #504
This was somewhat a red herring. The main issue was not the scratch,
but that the dynamic scheduling managed to run past the maximum
workset size under certain special circumstances.
If you do that, the scratch memory was not reallocated, because
that place had an independent check on validness.
The scratch memory access was the first memory access in the kernel though
and managed to crash then.
This should now be fixed.