Disable print of loaded profiling library
As discussed in kokkos/kokkos-tools#8,
this is one of the prints that really
annoys users at large MPI scale.
Since it is inside kokkos it won't be
affected by MPI-aware kokkos-tools.
This commit guards it with an #ifdef
that if now OFF by default, but can
be re-enabled in various ways in the
TriBITS and Makefile.kokkos systems.