Kokkos:  Add ability to control memory alignment threshold.
This commit adds the ability to set the memory alignment threshold, in a
manner similar to how it was done with the memory alignment.  Just
define KOKKOS_MEMORY_ALIGNMENT_THRESHOLD to the value you want to
override the default value of 4.  For example, set it to 0 to make sure
padding always happens.