Add test harness just for configuration
Because of the large number of permutations in both
architecture and devices, testing the header file and
build system files in an architecture-independent way is
useful for testing the development.

This is not hooked up to be run routinely but can be invoked with a:
  make -f UnitTestConfig.make
in the core/unit_test directory.