Improve cmake build system
The interface between the cmake build system and the
part of the makefile system for generating KokkosCore_config.h
and gen_kokkos.cmake is KOKKOS_SETTINGS.  The mapping from
cmake options to KOKKOS_SETTINGS is done in kokkos_options.cmake.
This is much clearer with improved diagnostics as well.
Also the configure_file that the previous cmake build system used
is removed as it is no longer needed.