Improve unit testing
Moving to a grep based system for the Makefile system to enable testing
just the setting of ARCH/DEVICES for builds which is the main goal,
and not worry about changing header files list will change the
makefiles.

Grepping does not work for the KokkosCore_config.h due to the complexity
of the mapping between KOKKOS_ARCH and the define variables.  Continue
to use diff tests for that.

Start on cmake testing similar to makefile testing as well.