Added testing for cmake configuration
Adding testing for cmake configuration in the same way as the
Makefile configuration.  This revealed some issues with the
consistency of ARCH/DEVICES among the various files, a bug in
the execute_command script, and some issues with Makefile testing.
By using two little shell scripts for doing the actual tests here,
it is much easier to do the actual testing for both systems so that is
used.