Enable TriBITS builds
Enabling TriBITS builds required separating out the part that is the
options (cmake/kokkos_options.cmake) and the setting of KOKKOS_SETTINGS
variable (cmake/kokkos_options.cmake) so that the TriBITS variables
could override the default kokkos variables.

As the top-level cmake was changed to enable TriBITS variables, some
advanced options that were not being set were discovered so some
additional options are now available in cmake builds.

Documentation on how to develop for both the cmake *and* Makefile builds
was changed from just explaining the architecture and devices to a more
general discussion of all the options.