Changes to make TriBITS builds compatible
This is meant to be compatible with the kokkos-config-207
branch of TriBITS and the trilinos-kokkos-1400 branch of
Trilinos.  The basic idea is to use Makefile.kokkos early
in the configuration (before packages) and then reuse that
information for the Kokkos build.  Much of the effort went
into making enabling backward compatibility.