Reduce calls to tracking_enabled even further
In talking with @dsunder we realized that the
rhs.m_record_bits should be checked prior to
calling tracking_enabled.
This change consolidates duplicate logic into
a macro that does this check first.
In theory, after this change, tracking_enabled
should never be called inside the user portion
of a Kokkos parallel algorithm.