DynamicView: Address issue #1206
Ccommit to modify DynamicView such that:
1. remove need for memory pool by...
2. resize only functions from the host outside of parallel_* regions
     reason - remove need for memory pool

Due to removing the memory pool, the ptr-array-to-chunks, i.e. m_chunks,
is now allocated in CudaUVM space when CudaSpace is enabled; chunks are
still allocated in CudaSpace.

Current impl API asks for (label, min_chunk_size, max_extent)
Max extent is asked for to avoid resizing the m_chunks array, which
could lead to possible issues if multiple DynamicViews are copies of
each other, sharing the allocation, and one of them requires resize of
the m_chunks array (to pointers of chunks)