Fix alignment needed for vector intrinsics
Vector intrinsics must be alligned, that means Kokkos views of
vector intrinsics must be alligned, even for a scalar view.
That means we need a real fix with a minimum allignment argument
to all the allocation calls in memory spaces. Until then, minimum
allignment is set to 64 for any length of view.