Runtime checking of View constructor args. Created cpp macro for KOKKOS_INVALID_INDEX. Added runtime checking for View constructors
Design point is that under new behavior, the number of dimensions passed
to a View constructor must exactly match the traits::rank_dynamic.

Added two macros, KOKKOS_CTOR_DEFAULT_ARG, which expands to 0 for old
behavior and to KOKKOS_INVALID_INDEX ( ~size_t(0) ) for the new behavior.

added separate check functions for host and device, so that a more
informative error can be printed on host.

The tests pass with and without --with-options=disable_deprecated_code
on ceelan.

The tests pass with and without --with-options=disable_deprecated_code
on white using
--with-cuda --compiler={KOKKOS_PATH}/bin/nvcc_wrapper
--with-cuda-options=enable_lambda --arch=Power8,Pascal60

test_all_sandia on white also passes.