ROCm: fix MDRange Reductions
the higher MDRangePolicy Ranks are using block_size larger 64.
The reduction was not working for it yet. The code copied from
CUDA didn't really work though, so I put some code in manually.
In most cases the number of wavefronts would anyway be only 1, 2
or 4. So doing the serial reduction should be fine instead of doing
a log algorithm.