(re)compile time fix for Kokkos on Darwin
recompilation times on Darwin were very long when building
kokkoskernels with kokkos because the `cp` command in the
Makefile for Kokkos was copying all the files into the KK
examples dir.  This updated the timestamps on the sources
each time resulting in a full recompile every time.

In the makefile, Darwin (OSX) systems did not use the `-u` parameter
to `cp` which makes `cp` only copy files over that are newer.  On
vanilla OSX systems, this parameter does not exist but for OSX
systems using homebrew to install the `coreutils` package, the
-u option is there and can be used to keep recompilation times on
darwin quick.