Change logic for determining iteration layout
when copying from two views in the same space
with different layouts.
Tests run on Power8 (Serial&OpenMP) and Kepler GPUs (Cuda) show that
the best-performing "layout" for the MDRange
iteration for this deep_copy variant is influenced
by much more than the destination view's layout,
and that the previous logic could produce a deep_copy
that was up to 3.5X slower than the alternative choice.
Using tests focused on rank-2 views, all the way from
square to 1M x 10, this complex set of rules was developed.

Also included in this commit is a simplification to
determine the execution space at compile time,
thus halving the number of instantiations of
DeepCopy that have to be done.