Refactor RangePolicy to reduce number of template instantiation.
RangePolicy is now a "template using alias" so that permutations of the
same template parameters do not instantiate different types.  The
underlying Impl::RangePolicy has been refactor to support constexpr
constructors and improved the number of integer ops to calculate the
WorkRange.

This also reduces the number of ParallelFor, ParallelReduce, and
ParallelScan instantiations.

Side Effect: Uses of 'RangePolicy' within the 'Kokkos::Impl' namespace
now needs to specify 'Kokkos::RangePolicy' instead of 'RangePolicy'.