Try to fix no-init mirror test
there was some confusion here about the difference
between allocating memory and initializing memory to zero.
the way this test works, the mirror will always have some
allocated memory associated with it.
the test was trying to assert that memory was not allocated
in some cases.
the intent was to check for lack of initialization to zero.
unfortunately, there is no way to assert non-initialization,
since we have no control over what the old garbage values where
(they could be all zero coincidentally).