Changes made after unit testing with miniMD benchmark:
update logic to apply the default path at the time the file is opened...not when the
dataspace is allocated.  This allows apps to change the default path between checkpoints.

update create_chkpt_mirror to print a warning and auto generate a view name if the src label
is blank.