adds support for more Kokkos::complex functions
Some notes on this commit:

o The previous implementation of Kokkos::sqrt did not match the stdc++
implementation, and resulted in Kokkos::sqrt(x*x) != x in some cases.
While the implementation details of the std::complex functions are not
defined by the C++ standard explicitly, I recommend this change to
prevent users being surprised by the above behavior.  The Kokkos::sqrt
behavior was discovered when trying to implement Kokkos::asinh, so I'm
committing these changes as an atomic commit for better clarity to
future Kokkos developers.

o The OpenMP Target backend tests fail for some of these functions.  The
cause seems to be orthogonal to these changes, however, so these new
tests are disabled for that particular backend (serial, openmp, and cuda
all pass and are enabled).  Engaging the tests for the OpenMP Target
backend will be addressed in a future PR.