Fixed -Wextra compiler warnings (Issue #3397)
The Sierra project is suffering from build errors when attempting to use
a recent Kokkos version to build trilinos/stk because we use the -Wextra
build flag and use warnings-as-errors.

Most of the warnings stem from mixing enum and integral types as possible
return values from nested ternary chains.  I investigated changing the
enum constants to "static constexpr size_t" constants, which should have
the same semantics and works in a Sierra build, but has link errors in
the Kokkos unit tests (due to unresolved symbol errors, likely because
of needing an external definition of the static which doesn't play
friendly in a header-only templated implementation).  The (ugly)
solution that always works is to sprinkle static_cast<size_t>()
everywhere.