Added support for SYCL indirect kernel execution.
An indirect kernel is one where we have a functor
that is not trivially copyable and so is explicitly
constructed by the host in USM shared memory before being passed
"by pointer" (inside a reference_wrapper) to SYCL parallel_for.

This is to address the limitation that SYCL
data types can only be implicitly copied to the device if they
are trivially copyable.