Cleaned up use of the __CUDACC__ macro
This messes with tooling because it's defined in the compiler
and not in a reachable header file. It should always be defined
if and only if KOKKOS_ENABLE_CUDA is defined, so I've added a
more helpful error message to that effect and removed it from
everywhere else in the code that it's used. We should probably
do something analogous for HIP in a different pull request.