In SYCLInternal, changed m_queue from a unique_ptr<sycl::queue> to an optional<sycl::queue>.
Ideally, it should just be a sycl::queue directly, but given
current issues with assignability under SYCL+CUDA, this is
the best we can do right now.

Because it was initially developed under C++14, unique_ptr
was used (as optional wasn't there), but it is better to
have fewer heap allocations, instead of unnecessary ones.