Fix false sharing in RandomNumber Pools
The pool arrays are accessed by all threads, but the
elements per thread are rather small and thus share cachelines.

Making the arrays 2D using the second dimension for padding only,
solves that problem. I have seen up to 200x improvement for
20 threads on skylake running the random number example with 100k and 1
as parameters. Serial performance is slightly reduced, but it only makes
the "grep a generator slower" by the equivalent of one double load more
and a few integer ops.