hpcbind: Use double quotes around $@ when invoking user command
At the end of the hpcbind script, when running the user command, the
command arguments are in the $@ variable. Unless $@ is enclosed by
double quotes, those arguments are subject to word splitting in bash,
which creates problems when those arguments have spaces that should be
preserved. An example that would fail without this commit is `hpcbind
-- foo args="hello world"`.