Intel 17 atomic illegal instruction fix
This one appears WITHOUT Desul atomics, so it wasn't related to them.
I.e. this one happened for a sync_fetch_add intrinsic which isn't
used in desul.

Otherwise similar symptoms as in #4223 (which we reverted in favor
of turning desul atomics off).
Only in this case it was on SKX not pre Haswell.