Improve m_copy_event handling
Kernels associated with functors which rely on IndirectKernelMemory
must wait for the memcpy to finish. Previously (prior to USM Pool),
this was implemented as a host-side fence on the copy operation.

A temporary solution was to have the copy event (m_copy_event) as a
public member of USMObjectMem, but this wasn't ideal. Now
SYCLFunctionWrapper has a method which returns the memcopy event, as
does USMObjectMem.

For trivially copyable kernels, the returned event is a
default-constructed one (which is immediately 'ready') and so won't
incur a wait.

Note that now the SYCLFunctionWrapper class has the associated
USMObjectMem as a member. As such, the 'Storage' argument to
register_event is probably superfluous & should probably be removed.