Demonstrate an error with Kokkos::sort
We have noticed that `Kokkos::sort` sometimes fails in HIP. This
adds a test that demonstrates this bug. When I compile Kokkos
on the crusher platform at OLCF using the ROCm 5.1.0 module, this
test fails. This can be used to determine what is going wrong
with sort and provide a regression test for it.