In vector::insert(iterator, InputIterator, InputIterator):
Changed the out of range check on it to use std::less, as one
cannot portably perform ordered pointer comparisons on pointers
that aren't pointing inside the same array (to a first approximation).

Removed the special case when inserting into an empty vector,
as it was incorrect (see issue #4975) and has no real performance
gain (unlike this special case in insert(it, count, val)).

Removed the special case when inserting zero elements, as there
is no need to hand optimize for this rare case, as it falls out
of the general algorithm anyway.

In vector::insert(it, count, val):

Changed the out of range check to use std::less

In TestVector.hpp:
Added tests for the above