put back binary operators with arithmetic types
It turns out the implicit conversion from arithmetic
types and SIMD+SIMD arithmetic operators is not sufficient
in the case of simd_abi::scalar where we have for example

operator+(simd<T, scalar>, simd<T, scalar>)

and we attempt to apply it to `simd<int, scalar>(1) + 1`

In this case the compiler is not able to deduce the parameter
T for parameter type simd<T,scalar> given the argument type `int`.

For this reason separate arithmetic operators are provided here
between SIMD types and arithmetic types, with numeric
promotion as needed also included (e.g. simd<int>*double=double)