Don't require user-defined volatile overloads in Kokkos::single
Prior to this commit, users would have to define the following
overloads (with unusual return types) to pass an aggregate to
Kokkos::single.
* KOKKOS_INLINE_FUNCTION void operator=(const volatile T &rhs)
  volatile {...}

This was annoying because it has to have a void return type to
prevent g++ from issuing a warning with the following usage:

T a, b;
a = b; // here

g++ will issue a warning if the return type was the expected `T&`
because `a=b` will have a type of `T&`, which results in a spurious
memory access if T is volatile (in g++).

it is also annoying because if the user wants to use T in any usual
value-type ways, they will have to follow the rule of 7 now that
they have defined one operator.