Make sure we don't add '-cuda' to the link line with NVC++
Obviously, in an ideal world we don't process that entire block when the
CUDA backend is not enabled.  Wrapping it in a big
if(KOKKOS_ENABLE_CUDA) makes the code less readable and I don't want to
get into refactoring the whole thing.  I considered introducing a
KOKKOS_CUDA_LINK_OPTIONS variables but that would force us to duplicate
some target_link_options workaround for CMake version less than 3.18