Allow as_view_of_rank_n() to be overloaded for "special" scalar types
as_view_of_rank_n() needs to be overloaded for special scalar types like
Sacado that need to pass a hidden dimension.  This change adds the
necessary SFINAE stuff to allow that to happen.  This will require
adding the new overload in Trilinos/Sacado.

For Trilinos issue #11018 causing runtime failures in Albany.