replace testc with movemask
testc was the wrong way to
compare masks because it
doesn't work if the second
mask has any false entries.
instead, use the movemask intrinsic
to convert a wide vector mask into
an int bitmask and compare that.
this should make the simd_mask::reference::operator bool
implementation a lot faster too