ensure both RDC options are consistent
the Kokkos_ENABLE_CUDA_RELOCATABLE_DEVICE_CODE
and CMAKE_CUDA_SEPARABLE_COMPILATION options
both control whether CUDA RDC is on when
we are building using CUDA as a CMake language.
This code makes sure that if either of those
options is on, the other one is also on.
This way, both older build scripts that used
the first option and "proper CMake" scripts
that use the latter will have the same effect
if building Kokkos with CUDA as a CMake
language is enabled.