Thread Safety: Call cudaSetDevice before almost any CUDA API call
when one creates a new thread, that thread will get the default
CUDA environment. I.e. any CUDA API call will initally be with
respect to device zero unless cudaSetDevice is called on that thread.

This means we have to call cudaSetDevice again before practically
any CUDA API call to be safe.

Currently the device id we use is a static variable. Subequently it
could be a per instance id. This would require us to change some of
these places, in particular the Kokkos::fence() impl which would
need to iterate over all active devices.