Add Kokkos::all_libs target (trilinos/Trilinos#11545)
This follows the guidance in "Professional CMake: 11th edition" section
"26.3. Installing Exports" to make sure there are now non-namespaced targets.
The idiom is not pretty but that is what you are supposed to do, and it works.

This creates a Kokkos::all_libs targets in the current project and puts an
imported interface Kokkos::all_libs target in the KokkosTargets.cmake file.