Kokkos: Mark HWLOC as a TriBITS TPL as well (trilinos/Trilinos#11938)
It was reported by a Trilinos user in trilinos/Trilinos#11938 that after
removing Kokkos subpackages that their build of Trilinos failed and by marking
'HWLOC' as a TriBITS TPL this fixed the problem.  HWLOC has always been a
TriBITS TPL so it is not quite clear how this worked before this.