support vector_aligned_tag
this SIMD load/store tag is defined in
the ISO C++ TS and specifies that the
load or store operation is being done
with a pointer aligned to vector width.
The Intel backends (AVX2, AVX512) do
have different intrinsics in this case,
and this enhancement allows users of
Kokkos SIMD to make use of those intrinsics.