Remove calling tribits_exclude_autotools_files()
Kokkos has no autotools files anymore and support for
tribits_exclude_autotools_files() was stripped out of TriBITS because it is
not needed anymore.

There will be a slight hack in Trilinos 'develop' to avoid needing to make
this change on Trilinos 'develop'.  But the next time Kokkos is snapshotted
into Trilinos, we can remove that (minor) hack.