Prefer defaulted default constructor for Bitset
The default argument to the constructor that takes the size of the
bitset was deferring to another constructor that creates an empty view
with a label argument.  This alocates 128 bits for the view header.
This showed when constructing an UnorderedMap with a pointless 128-bit
"header-only" allocation which implies an unnecessary fence.