Make Host backends be included before device backends
There was an inclusion order issue when trying to enable
unified memory, which was due to incomplete types.
Effectively the Host backends must be defined before
the device memory spaces for them to be marked accessible.

We may have the same issue if we try to mark HostSpace accessible
by device memory spaces - but I am not sure.