Accommodate users that depend on a code that define silly macros
Reported on Slack, user was getting when including some other dependency
```
<prefix>/include/Kokkos_Tuners.hpp(259): error: expected a "," or ">"
 template <typename PointType, size_t 0>
                    ^
```
It turns out that dependency defines macros such as "X", "Y", "Z", or
even "DIM" which is really asking for trouble.  See here
https://github.com/fluiddynsci/EngSketchPad/blob/1fe3fc4c68a759e0832c02f8d3c2bd8722f183a8/include/libCart3D/c3d_global.h#L24-L28

Since it does not take much change to accommodate that user, we avoid
named parameters that collide with these silly macro defines.
We are not promising that we support it but this should help.