Deprecate specialization of Kokkos::pair for a single element
This specialization is not documented, does not follow the standard
library, it is not tested and has no known usage in Trilinos.
`Kokkos::pair`, as we generally describe it, was intended as a drop-in
replacement for `std::pair`.  Hence, obscure departure from the standard
implementation do not look like a good idea.
This PR suggest to deprecate that `T2=void` specialization for
degenerate pair that only hold one element.